fix(api-headless-cms): rename adstractions folder to abstractions
