+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IDeleteEntry.ts
