+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IDeleteEntryOperation.ts
