+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IGetEntriesByIds.ts
