+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IGetEntry.ts
