+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IGetLatestRevisionByEntryId.ts
