+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IGetPreviousRevisionByEntryId.ts
