+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IGetPublishedRevisionByEntryId.ts
