+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IGetRevisionById.ts
