+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IGetRevisionsByEntryId.ts
