+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IListEntries.ts
