+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IListEntriesOperation.ts
