+++ packages/api-headless-cms/src/crud/contentEntry/abstractions/IMoveEntryToBinOperation.ts
