+++ packages/api-headless-cms/src/crud/contentEntry/useCases/DeleteEntry/MoveEntryToBinOperation.ts
@@ -1,4 +1,4 @@
-import { IMoveEntryToBinOperation } from "~/crud/contentEntry/adstractions";
+import { IMoveEntryToBinOperation } from "~/crud/contentEntry/abstractions";
 import {
     CmsEntryStorageOperations,
     CmsEntryStorageOperationsMoveToBinParams,
