+++ packages/api-headless-cms/src/crud/contentEntry/useCases/DeleteEntry/MoveEntryToBinOperationWithEvents.ts
@@ -1,5 +1,5 @@
 import WebinyError from "@webiny/error";
-import { IMoveEntryToBinOperation } from "../../adstractions";
+import { IMoveEntryToBinOperation } from "../../abstractions";
 import { CmsEntryStorageOperationsMoveToBinParams, CmsModel } from "~/types";
 import { DeleteEntryUseCasesTopics } from "./index";
