+++ packages/api-headless-cms/src/crud/contentEntry/useCases/GetEntriesByIds/GetEntriesByIds.ts
@@ -1,4 +1,4 @@
-import { IGetEntriesByIds } from "../../adstractions";
+import { IGetEntriesByIds } from "../../abstractions";
 import {
     CmsEntryStorageOperations,
     CmsEntryStorageOperationsGetByIdsParams,
