+++ packages/api-headless-cms/src/crud/contentEntry/useCases/GetLatestEntriesByIds/GetLatestEntriesByIdsSecure.ts
@@ -1,6 +1,6 @@
 import { AccessControl } from "~/crud/AccessControl/AccessControl";
 import { filterAsync } from "~/utils/filterAsync";
-import { IGetLatestEntriesByIds } from "../../adstractions";
+import { IGetLatestEntriesByIds } from "../../abstractions";
 import { CmsEntryStorageOperationsGetLatestByIdsParams, CmsModel } from "~/types";
 
 export class GetLatestEntriesByIdsSecure implements IGetLatestEntriesByIds {
