+++ packages/api-headless-cms/src/crud/contentEntry/useCases/GetRevisionsByEntryId/GetRevisionsByEntryId.ts
@@ -1,4 +1,4 @@
-import { IGetRevisionsByEntryId } from "../../adstractions";
+import { IGetRevisionsByEntryId } from "../../abstractions";
 import {
     CmsEntryStorageOperations,
     CmsEntryStorageOperationsGetRevisionParams,
