+++ packages/api-headless-cms/src/crud/contentEntry/useCases/ListEntries/GetEntry.ts
@@ -1,4 +1,4 @@
-import { IGetEntry, IListEntriesOperation } from "~/crud/contentEntry/adstractions";
+import { IGetEntry, IListEntriesOperation } from "~/crud/contentEntry/abstractions";
 import { CmsEntryGetParams, CmsModel } from "~/types";
 import { NotFoundError } from "@webiny/handler-graphql";
