+++ packages/api-headless-cms/src/crud/contentEntry/useCases/ListEntries/ListEntriesOperationNotDeleted.ts
@@ -1,4 +1,4 @@
-import { IListEntriesOperation } from "../../adstractions";
+import { IListEntriesOperation } from "../../abstractions";
 import { CmsEntryStorageOperationsListParams, CmsModel } from "~/types";
 
 export class ListEntriesOperationNotDeleted implements IListEntriesOperation {
