+++ packages/api-headless-cms/src/crud/contentEntry/useCases/ListEntries/ListEntriesOperationWithSort.ts
@@ -1,4 +1,4 @@
-import { IListEntriesOperation } from "../../adstractions";
+import { IListEntriesOperation } from "../../abstractions";
 import { CmsEntryListSort, CmsEntryStorageOperationsListParams, CmsModel } from "~/types";
 
 export class ListEntriesOperationWithSort implements IListEntriesOperation {
